Remove IRequestPipeline
